+++ src/models/guildrank.ts
@@ -1,0 +1,8 @@
+export enum GuildRank {
+    NoRank = -1,
+    Initiate = 0,
+    Member = 10,
+    Officer = 20,
+    Leader = 30,
+    Founder = 40
+}